MinGraph-UNet/experiments/ablation_study.py
# experiments/ablation_study.py
import yaml
import os
import argparse

# Import evaluation functions from other experiment scripts or metrics.py
from experiments.yield_estimation_performance import evaluate_yield_model 
# Or from segmentation_performance if ablation is on seg metrics

def load_config(config_dir, config_name):
    with open(os.path.join(config_dir, config_name), 'r') as f:
        return yaml.safe_load(f)

def run_ablation_experiment(config_path, ablation_variant_name, model_weights_path):
    """
    Runs evaluation for a specific ablation study variant.
    """
    print(f"\n--- Running Ablation Study for Variant: {ablation_variant_name} ---")
    print(f"Using weights: {model_weights_path}")

    # The core challenge is that `model_type` in evaluate_yield_model needs to correctly
    # instantiate or represent the ablated model.
    # E.g., if variant_name is "Min-Cut Only", the loaded model should only use Min-Cut logic.
    # This might require modifications to evaluate_yield_model or a more general eval function.

    # For simplicity, let's assume `evaluate_yield_model` can handle a generic "model_type"
    # that corresponds to an ablated version.
    # Or, each ablation variant is treated as a distinct `model_type`.

    # Table 3 has "Count Accuracy" and "Error", so it's yield-focused.
    # The `model_type` passed to evaluate_yield_model would be the `ablation_variant_name`
    # or a mapping from it to a loadable model configuration.
    
    # This is a conceptual mapping. The `evaluate_yield_model` would need to
    # correctly instantiate the model based on `ablation_variant_name`.
    # E.g., if "Min-Cut Only", it loads a model that only does U-Net -> PatchGAT -> MinCut-Predictor -> Count.
    
    # If your `train_end_to_end.py` can be configured to train these ablated versions,
    # then `evaluate_yield_model` (with a flexible model loader) could work.
    # For now, let's assume model_type matches ablation_variant_name for placeholder.
    
    # The `model_weights_path` must point to the weights of this specific ablated model.
    if not os.path.exists(model_weights_path):
        print(f"Warning: Weights for {ablation_variant_name} not found at {model_weights_path}. Using mock evaluation.")
        # Simulate some results for placeholder
        mock_results = {
            'count_accuracy_perc': np.random.uniform(80,95),
            'yield_estimation_error_perc': np.random.uniform(5,15)
        }
        print(f"Mock Results - Count Accuracy: {mock_results['count_accuracy_perc']:.2f}%, Error: {mock_results['yield_estimation_error_perc']:.2f}%")
        return mock_results['count_accuracy_perc'], mock_results['yield_estimation_error_perc']


    # Assuming `evaluate_yield_model` can take `ablation_variant_name` as model_type
    # and load/configure the model accordingly. This is a strong assumption.
    # In reality, you might have different scripts or model loading logic for each variant.
    try:
        results = evaluate_yield_model(config_path, model_type=ablation_variant_name, model_weights_path=model_weights_path)
        count_acc = results['count_accuracy_perc']
        error_perc = results['yield_estimation_error_perc']
        print(f"Results for {ablation_variant_name} - Count Accuracy: {count_acc:.2f}%, Error: {error_perc:.2f}%")
        return count_acc, error_perc
    except Exception as e:
        print(f"Error evaluating ablation variant {ablation_variant_name}: {e}")
        # Return dummy failing values
        return -1, -1


if __name__ == '__main__':
    parser = argparse.ArgumentParser(description="Run Ablation Study.")
    parser.add_argument('--config_path', type=str, default="configs/", help="Path to config files dir.")
    # Individual model weights for each variant would be needed.
    # For a full run, you'd pass paths to these weights.
    # Example: --mincut_only_weights /path/to/mincut_only.pth ...

    args = parser.parse_args()

    ablation_variants = {
        # Variant Name from Table 3 : Path to its trained weights
        "Min-Cut Only": "path/to/weights_min_cut_only.pth",
        "Graph U-Net Only": "path/to/weights_graph_unet_only.pth", # U-Net + GATs
        "Graph Construction": "path/to/weights_graph_construction_variant.pth", # (This name is vague in table, implies importance of how graph is built)
        "Graph Traversal": "path/to/weights_graph_traversal_variant.pth", # (Implies importance of GATs themselves)
        "Combined (Full Method)": "path/to/weights_full_method.pth"
    }

    print("--- Starting Ablation Study Evaluation (Yield Metrics) ---")
    print("Variant Name               | Count Acc. (%) | Error (%)")
    print("---------------------------|----------------|----------")

    for variant_name, weights_path in ablation_variants.items():
        # This is where you'd ensure the correct model variant is loaded and evaluated.
        # The `evaluate_yield_model` would need to be adapted to instantiate
        # these specific ablated model architectures based on `variant_name`.
        # For this example, we rely on the placeholder behavior if weights don't exist.
        
        # Create dummy weight files if they don't exist for the script to run
        if not os.path.exists(os.path.dirname(weights_path)):
            try: os.makedirs(os.path.dirname(weights_path))
            except: pass # Might fail if path is just a filename
        if not os.path.exists(weights_path):
            with open(weights_path, 'w') as wf: wf.write("dummy ablation weights")
            print(f"Created dummy weights for {variant_name} at {weights_path}")

        count_acc, error_perc = run_ablation_experiment(args.config_path, variant_name, weights_path)
        print(f"{variant_name:<26} | {count_acc:14.2f} | {error_perc:8.2f}")

        # Clean up dummy weights if created by this script run
        if os.path.exists(weights_path) and "path/to/" in weights_path : # Simple check if it's a dummy path
            try: os.remove(weights_path)
            except: pass


    # Dummy configs for ablation (similar to yield_estimation_performance)
    if not os.path.exists(args.config_path): os.makedirs(args.config_path)
    with open(os.path.join(args.config_path, "dataset.yaml"), 'w') as f:
        yaml.dump({'data_root': 'temp_mango_dataset_ablation', 'test_dir': 'test', 'image_folder':'images'}, f)
    with open(os.path.join(args.config_path, "model.yaml"), 'w') as f: yaml.dump({},f)
    with open(os.path.join(args.config_path, "preprocessing.yaml"), 'w') as f:
        yaml.dump({'resize_dim': [128,128], 'normalization_mean': [0.485,0.456,0.406], 'normalization_std': [0.229,0.224,0.225]},f)
    with open(os.path.join(args.config_path, "training.yaml"), 'w') as f:
         yaml.dump({'device':'cpu', 'eval_batch_size': 2, 'num_workers':0},f)